+++ init_graph.py
@@ -1,33 +1,15 @@
-
-    def TrainGraph(session_in):
-        for i in range(1001):
-            batch = mnist.train.next_batch(batchSize)
-            session_in.run(
-                train_step, feed_dict={x: batch[0],
-                                       true_y: batch[1],
-                                       keep_prob: 0.5})
-            if i % 100 == 0 and i != 0:
-                trainAccuracy = session_in.run(
-                    accuracy, feed_dict={x: batch[0],
-                                         true_y: batch[1],
-                                         keep_prob: 1.0})
-                print("step %d, training accuracy %g" % (i, trainAccuracy))
+import tensorflow as tf
+import tensorflow.contrib.slim as slim
+from tensorflow.examples.tutorials.mnist import input_data
+import math
+import session as loc_sess
 
 
-    def TestGraph(session_in):
-        testAccuracy = session_in.run(
-            accuracy,
-            feed_dict={x: mnist.test.images,
-                       true_y: mnist.test.labels,
-                       keep_prob: 1.0})
-        print("test accuracy %g" % (testAccuracy))
+def InitGraph(mode):
+    print("init graph...")
+    if mode in ("-f", "--fetch"):
+        loc_sess.Fetch(mode)
 
-
-        def ResetGraph(mode):
-
-        if(mode)
-            mnist = input_data.read_data_sets("MNIST_data/", one_hot=True)
-            
         tf.reset_default_graph()
 
         x = tf.placeholder(tf.float32, [None, 784], name="x-in")
@@ -55,4 +37,4 @@
         init = tf.global_variables_initializer()
         sess.run(init)
 
-        TrainGraph(sess)
+    #    TrainGraph(sess)
